switch to branch: handle slashes in branch names
In git, having slashes in branch names are not uncommon.

We cannot use a branch name with slashes in the url, so in that case, we
use the changeset id instead like we already do for tags and bookmarks.

Fixes #654

--HG--
branch : beta
extra : source : 60c63fcc91e37e4f79fd4da8e94acb7d6ed7ce76