html: move "Submit a bug" to make it more clear that it is for RhodeCode, not the repo
RhodeCode _could_ contain a bug tracker and this link _could_ be for filing
bugs for the hosted projects.

Moving the link to the RhodeCode info makes it more clear that it is for
RhodeCode bugs.

The server instance is however something local, not directly related to the
upstream.

--HG--
branch : beta