pullrequests: fix changesets ordering being reversed when creating new pull requests
41b4edf77b5b tried to make the displayed order of changeset consistent: The
topmost is always the latest.

That did however also reverse the ordering of the changesets sent back in the
post when used in the pull request creation form. Displaying the pull request
later on would reverse it again and thus show it in the 'wrong' order.

We now undo that reversing when creating the pull requests, and the stored data
will thus be the same as before.

--HG--
branch : beta
extra : source : ef13f741f12f10db8236a4e76ba1a4e89d2a875c